Test the env var sourcing -> decimals test should pass

+++ test/src/test/java/org/hiero/mirror/test/e2e/acceptance/client/AccountClient.java
@@ -61,6 +61,11 @@ public class AccountClient extends AbstractNetworkClient {
 
     @Override
     public void clean() {
+        if (Boolean.parseBoolean(System.getenv("CI"))) {
+            // In CI we don't want to cleanup as the entities are needed in the k6 test in the next step.
+            log.warn("Acceptance tests running in CI -> skip cleanup.");
+            return;
+        }
         log.info("Deleting {} accounts", accountIds.size());
         deleteAll(accountIds, this::delete);
